Refactor service deployment info resource and broker
Fix tests
